Add BPF_J{LT,LE,SLT,SLE} instructions
See kernel patches merged in commit 078295fb9af5. Those new instructions
were added for performance reason to the kernel.

Add them to rbpf so as to remain as compatible as possible with Linux.

This commit adds the new instructions for the interpreter, the JIT, the
verifier, the assembler, the disassembler and the program builder. Unit
tests are also added for these instructions.

Fixes issue #18.