Add support for dynamic stack frames (#274)
This change implements dynamic stack frames when config.dynamic_stack_frames is
set to true.

When dynamic_stack_frames=true, r11 is exposed as the stack pointer. The stack is
fully descending, so `sub r11, N` can be used to grow it, and `add r11, N` to
shrink it. Those are the only operations allowed on r11, which are special cased
by the interpreter and jit to resize the stack. No other operations are allowed
on the stack register.