fixup: Avoid host build of SBPF program test crates (#1711) (#2145)
In

  commit d5cd2c4
  Author: Alexander Meißner <AlexanderMeissner@gmx.net>
  Date:   Tue Jun 18 23:10:40 2024 +0200

      Refactor - Avoid host build of SBPF program test crates (#1711)

we changed the test selection process slightly, switching from the
`test-bpf` feature to `sbf_rust` feature, and marking individual tests,
rather than the whole module.  The problem is that now when I run

    cd program/sbf
   ../../cargo check --bins --tests

I see unused warnings for all the modules imported by the `simulation`
and `sysvar` modules.  At the later are, effectively, empty when
`sbf_rust` is not set.

Moving the config flag back to the module level seems to do the intended
thing and removes the warnings.  Alternatively, we could mark the `use`
statement with an explicit `#[cfg(feature = "sbf_rust")]`, but it seems
that all the tests are designed to be run only when `sbf_rust` is set.